updating componentS
